Add DNS fallback IP to be used before dnscrypt starts
If no fallback IP is set then on server startup dnscrypt-proxy is unable to get the IP addresses of the DNS servers to use. Added the Cloudflare 1.1.1.1 DNS server as the default fallback when the pihole configuration option is set; once dnscrypt-proxy is able to download the DNS server list, it uses the servers in that list and no longer uses the DNS fallback.